+++ docs/source/conf.py
@@ -21,7 +21,7 @@ author = "Parker Hancock"
 
 # The full version, including alpha/beta/rc tags
 
-version = release = "1.1.3"
+version = release = "1.1.4"