+++ src/patent_client/itc_edis.py
@@ -6,10 +6,11 @@ import time
 import xml.etree.ElementTree as ET
 
 import requests
-from patent_client import CACHE_BASE
-from patent_client import SETTINGS
 
-from .util import Manager, Model, one_to_many, one_to_one
+from patent_client import SETTINGS
+from patent_client import session
+
+from .util import IterableManager, Model, one_to_many, one_to_one
 
 CLIENT_SETTINGS = SETTINGS["ItcEdis"]
 if os.environ.get("EDIS_USER", False):
@@ -20,15 +21,15 @@ else:
     PASSWORD = CLIENT_SETTINGS["Password"]
 
 BASE_URL = "https://edis.usitc.gov/data"
-CACHE_DIR = CACHE_BASE / "itc_edis"
-CACHE_DIR.mkdir(exist_ok=True)
-
-session = requests.Session()
 
 # API Guide at: https://www.usitc.gov/docket_services/documents/EDIS3WebServiceGuide.pdf
 
+SECRET_KEY = None
 
-class ITCInvestigationManager(Manager):
+class AuthenticationException(Exception):
+    pass
+
+class ITCInvestigationManager(IterableManager):
     max_retries = 3
     auth_time = 10 * 60  # Re-authenticate every # seconds
     last_auth = 0
@@ -40,36 +41,33 @@ class ITCInvestigationManager(Manager):
             > ITCInvestigationManager.auth_time
         ):
             path = "/secretKey/" + USERNAME
-            response = session.get(BASE_URL + path, params={"password": PASSWORD})
+            with session.cache_disabled():
+                response = session.get(BASE_URL + path, params={"password": PASSWORD})
+            if not response.ok:
+                import pdb; pdb.set_trace()
+                raise AuthenticationException("EDIS Authentication Failed! Did you provide the correct username and password?")
             tree = ET.fromstring(response.text)
-            key = tree.find("secretKey").text
-            session.auth = (USERNAME, key)
+            SECRET_KEY = tree.find("secretKey").text
 
     def filter(self):
         raise NotImplementedError("EDIS Api does not have a search function!")
 
     def get(self, investigation_number):
-        fname = CACHE_DIR / (investigation_number + ".json")
-        if fname.exists():
-            return ITCInvestigation(json.load(open(fname)))
-        else:
-            ITCInvestigationManager.authenticate()
-            url = self.base_url + investigation_number
-            response = session.get(url)
-            tree = ET.fromstring(response.text)
-            tree = tree[0][0]
-            data = {
-                "phase": tree.find("investigationPhase").text,
-                "number": tree.find("investigationNumber").text,
-                "status": tree.find("investigationStatus").text,
-                "title": tree.find("investigationTitle").text,
-                "type": tree.find("investigationType").text,
-                "doc_list_url": tree.find("documentListUri").text,
-                "docket_number": tree.find("docketNumber").text,
-            }
-            with open(fname, "w") as f:
-                json.dump(data, f, indent=2)
-            return ITCInvestigation(data)
+        ITCInvestigationManager.authenticate()
+        url = self.base_url + investigation_number
+        response = session.get(url, auth=(USERNAME, SECRET_KEY))
+        tree = ET.fromstring(response.text)
+        tree = tree[0][0]
+        data = {
+            "phase": tree.find("investigationPhase").text,
+            "number": tree.find("investigationNumber").text,
+            "status": tree.find("investigationStatus").text,
+            "title": tree.find("investigationTitle").text,
+            "type": tree.find("investigationType").text,
+            "doc_list_url": tree.find("documentListUri").text,
+            "docket_number": tree.find("docketNumber").text,
+        }
+        return ITCInvestigation(data)
 
 
 class ITCInvestigation(Model):
@@ -80,7 +78,7 @@ class ITCInvestigation(Model):
         return f"<ITCInvestigation(number={self.number})>"
 
 
-class ITCDocumentsManager(Manager):
+class ITCDocumentsManager(IterableManager):
     primary_key = "investigation_number"
     base_url = BASE_URL + "/document"
     allowed_filters = {
@@ -92,17 +90,11 @@ class ITCDocumentsManager(Manager):
     }
 
     def get(self, document_id):
-        fname = CACHE_DIR / f"document-{document_id}.json"
-        if not fname.exists():
-            ITCInvestigationManager.authenticate()
-            response = session.get(f"{self.base_url}/{document_id}")
-            tree = ET.fromstring(response.text)
-            doc_el = tree.find(".//document")
-            data = self.parse_doc(doc_el)
-            with open(fname, "w") as f:
-                json.dump(data, f, indent=2)
-        else:
-            data = json.load(open(fname))
+        ITCInvestigationManager.authenticate()
+        response = session.get(f"{self.base_url}/{document_id}", auth=(USERNAME, SECRET_KEY))
+        tree = ET.fromstring(response.text)
+        doc_el = tree.find(".//document")
+        data = self.parse_doc(doc_el)
         return ITCDocument(data)
 
     def parse_doc(self, element):
@@ -126,25 +118,21 @@ class ITCDocumentsManager(Manager):
             data[key] = element.find(value).text
         return data
 
-    def get_item(self, key):
-        query = {self.allowed_filters[k]: v for (k, v) in self.filter_params.items()}
-        page = int(key / 100) + 1
-        location = key % 100
-        query["pagenumber"] = page
-        q_string = re.sub(r'[\{\}":, ]+', "-", json.dumps(query, sort_keys=True)[1:-1])
-        fname = CACHE_DIR / f"document-page-{q_string}.json"
-        if fname.exists():
-            data = json.load(open(fname))
-        else:
+    def __iter__(self):
+        query = {self.allowed_filters[k]: v for (k, v) in self.config['filter'].items()}
+        page = 1
+        page_length = None
+        while not page_length or page_length >= 100:
+            query["pagenumber"] = page
+            q_string = re.sub(r'[\{\}":, ]+', "-", json.dumps(query, sort_keys=True)[1:-1])
             ITCInvestigationManager.authenticate()
-            response = session.get(self.base_url, params=query)
+            response = session.get(self.base_url, params=query, auth=(USERNAME, SECRET_KEY))
             tree = ET.fromstring(response.text)[0]
-            data = list()
-            for element in tree.findall("document"):
-                data.append(self.parse_doc(element))
-            with open(fname, "w") as f:
-                json.dump(data, f, indent=2)
-        return ITCDocument(data[location])
+            page_length = len(tree.findall("document"))
+            docs = map(self.parse_doc, tree.findall("document"))
+            for doc in docs:
+                yield ITCDocument(doc)
+            page += 1
 
 
 class ITCDocument(Model):
@@ -158,40 +146,32 @@ class ITCDocument(Model):
         return f"<ITCDocument(title={self.title})>"
 
 
-class ITCAttachmentManager(Manager):
+class ITCAttachmentManager(IterableManager):
     primary_key = "investigation_number"
     base_url = BASE_URL + "/attachment/"
     allowed_filters = ["document_id"]
 
-    def get_item(self, key):
-        doc_id = self.filter_params["document_id"]
-        fname = CACHE_DIR / f"attachments-{doc_id}.json"
-        if fname.exists() and False:
-            data = json.load(open(fname))
-        else:
-            ITCInvestigationManager.authenticate()
-            response = session.get(self.base_url + doc_id)
-            tree = ET.fromstring(response.text)
-            attribute_dict = dict(
-                id="id",
-                document_id="documentId",
-                title="title",
-                size="fileSize",
-                file_name="originalFileName",
-                pages="pageCount",
-                created_date="createDate",
-                last_modified_date="lastModifiedDate",
-                download_url="downloadUri",
-            )
-            data = list()
-            for element in tree.findall(".//attachment"):
-                row = dict()
-                for k, value in attribute_dict.items():
-                    row[k] = element.find(value).text
-                data.append(row)
-            with open(fname, "w") as f:
-                json.dump(data, f, indent=2)
-        return ITCAttachment(data[key])
+    def __iter__(self):
+        doc_id = self.config['filter']["document_id"]
+        ITCInvestigationManager.authenticate()
+        response = session.get(self.base_url + doc_id, auth=(USERNAME, SECRET_KEY))
+        tree = ET.fromstring(response.text)
+        attribute_dict = dict(
+            id="id",
+            document_id="documentId",
+            title="title",
+            size="fileSize",
+            file_name="originalFileName",
+            pages="pageCount",
+            created_date="createDate",
+            last_modified_date="lastModifiedDate",
+            download_url="downloadUri",
+        )
+        for element in tree.findall(".//attachment"):
+            row = dict()
+            for k, value in attribute_dict.items():
+                row[k] = element.find(value).text
+            yield ITCAttachment(row)
 
 
 class ITCAttachment(Model):
@@ -204,13 +184,10 @@ class ITCAttachment(Model):
     def download(self, path="."):
         *_, ext = self.file_name.split(".")
         filename = f"{self.document.title.strip()} - {self.title}.{ext}"
-        cdir = os.path.join(CACHE_DIR, self.document.investigation.number)
-        os.makedirs(cdir, exist_ok=True)
-        cname = os.path.join(cdir, filename)
         oname = os.path.join(path, filename)
-        if not os.path.exists(cname):
-            response = session.get(self.download_url, stream=True)
-            with open(cname, "wb") as f:
+        if not os.path.exists(oname):
+            response = session.get(self.download_url, auth=(USERNAME, SECRET_KEY), stream=True)
+            with open(oname, "wb") as f:
                 for chunk in response.iter_content(1024):
                     f.write(chunk)
-        shutil.copy(cname, oname)
+        return oname