+++ tests/test_itc_edis.py
@@ -4,7 +4,7 @@ from patent_client import ITCInvestigation
 
 import pytest
 
-@pytest.mark.skip('Due to the government shutdown, EDIS is offline')
+@pytest.mark.skip("Authentication Problem")
 class TestItcEdis:
     def test_can_get_investigation(self):
         inv = ITCInvestigation.objects.get("337-TA-1025")
@@ -20,8 +20,8 @@ class TestItcEdis:
 
     def test_can_get_documents(self):
         inv = ITCInvestigation.objects.get("337-TA-966")
-        print(inv.documents[0].data)
-        assert inv.documents[0].data == {
+        print(inv.documents.first().data)
+        assert inv.documents.first().data == {
             "investigation_number": "337-966",
             "type": "Voting Sheet",
             "title": None,
@@ -36,11 +36,13 @@ class TestItcEdis:
             "last_modified": "2017/05/23 15:43:34",
             "id": "600197",
         }
-        assert inv.documents[0].investigation.number == "337-966"
+        assert inv.documents.first().investigation.number == "337-966"
 
     def test_can_get_attachments(self, tmpdir):
         inv = ITCInvestigation.objects.get("337-TA-1025")
-        assert inv.documents[15].attachments[0].data == {
+        doc_iterator = iter(inv.documents)
+        docs = [next(doc_iterator) for i in range(16)]
+        assert docs[15].attachments.first().data == {
             "id": "1163223",
             "document_id": "605365",
             "title": "605365",
@@ -51,7 +53,7 @@ class TestItcEdis:
             "last_modified_date": datetime.date(2017, 3, 13),
             "download_url": "https://edis.usitc.gov/data/download/605365/1163223",
         }
-        inv.documents[15].attachments[0].download(tmpdir)
+        docs[15].attachments.first().download(tmpdir)
         expected_doc = (
             tmpdir
             / "Granting Respondent's Unopposed Motion for Leave to Designate a Substitute Expert - 605365.pdf"